requirements.txt
# Fuzzy Attention Networks - Requirements
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Web interface
streamlit>=1.28.0
plotly>=5.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.7